Covrr-Push-Api/pyproject.toml
[tool.poetry]
name = "jimi-push-api"
version = "0.1.0"
description = ""
authors = ["shashi <shashi@infi.money>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.7.1"
redis = "^5.2.1"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
